+++ lib/number_display.dart
@@ -3,14 +3,54 @@ library number_display;
 
 import 'dart:math';
 
+const maxDoubleLength = 16;
+
+enum RoundingType {
+  round,
+  floor,
+  ceil,
+}
+
+List<String> _rounding(String intStr, String decimalStr, int decimalLength, RoundingType type) {
+  intStr = intStr ?? '';
+  if ((decimalStr == null) || (decimalStr == '0')) {
+    decimalStr = '';
+  }
+  if (decimalStr.length <= decimalLength) {
+    return [intStr, decimalStr];
+  }
+  if (intStr.length + decimalLength > maxDoubleLength) {
+    return [intStr, decimalStr.substring(0, min(decimalLength, decimalStr.length))];
+  }
+  decimalLength = max(decimalLength, 0);
+  final value = double.parse('$intStr.${decimalStr}e$decimalLength');
+  List<String> rstStrs;
+  if (type == RoundingType.ceil) {
+    rstStrs = (value.ceil() / pow(10, decimalLength)).toString().split('.');
+  } else if (type == RoundingType.floor) {
+    rstStrs = (value.floor() / pow(10, decimalLength)).toString().split('.');
+  } else {
+    rstStrs = (value.round() / pow(10, decimalLength)).toString().split('.');
+  }
+  if (rstStrs.length == 2) {
+    if (rstStrs[1] == '0') {
+      rstStrs[1] = '';
+    }
+    return rstStrs;
+  }
+  return [rstStrs[0], ''];
+}
+
 typedef Display = String Function(num value);
 
 Display createDisplay({
   int length = 9,
-  int decimal = 2,
+  int precision,
   String placeholder = '',
-  bool comma = true,
+  bool separator = true,
+  RoundingType roundingType = RoundingType.round,
 }) => (num value) {
+  precision ??= length;
   placeholder = placeholder.substring(0, min(length, placeholder.length));
 
   if (value == null || !value.isFinite) {
@@ -19,30 +59,32 @@ Display createDisplay({
 
   final valueStr = value.toString();
   final negative = RegExp(r'^-?').stringMatch(valueStr) ?? '';
-  final integer = RegExp(r'\d+').stringMatch(valueStr) ?? '';
+
+  var roundingRst = _rounding(
+    RegExp(r'\d+').stringMatch(valueStr) ?? '',
+    RegExp(r'(?<=\.)\d+$').stringMatch(valueStr) ?? '',
+    precision,
+    roundingType,
+  );
+  var integer = roundingRst[0];
+  var deci = roundingRst[1];
   final localeInt = integer.replaceAllMapped(
     RegExp(r'(\d{1,3})(?=(\d{3})+(?!\d))'),
     (m) => '${m[1]},',
   );
-  final deciRaw = RegExp(r'(?<=\.)\d+$').stringMatch(valueStr) ?? '';
-  final deci = deciRaw
-    .substring(0, min(decimal, deciRaw.length))
-    .replaceAll(RegExp(r'0+$'), '');
 
   int currentLen = negative.length + localeInt.length + 1 + deci.length;
-  String deciShow;
-  if (comma && currentLen <= length) {
-    deciShow = deci.replaceAll(RegExp(r'0+$'), '');
-    return '${negative}${localeInt}${deciShow == '' ? '' : '.'}${deciShow}';
+  if (separator && currentLen <= length) {
+    deci = deci.replaceAll(RegExp(r'0+$'), '');
+    return '${negative}${localeInt}${deci == '' ? '' : '.'}${deci}';
   }
 
   int space = length - negative.length - integer.length;
-  if (space >= 2) {
-    deciShow = deci.substring(0, min(space - 1, deci.length)).replaceAll(RegExp(r'0+$'), '');
-    return '${negative}${integer}${deciShow == '' ? '' : '.'}${deciShow}';
-  }
   if (space >= 0) {
-    return '${negative}${integer}';
+    roundingRst = _rounding(integer, deci, space - 1, roundingType);
+    integer = roundingRst[0];
+    deci = roundingRst[1].replaceAll(RegExp(r'0+$'), '');
+    return '${negative}${integer}${deci == '' ? '' : '.'}${deci}';
   }
 
   final sections = localeInt.split(',');
@@ -52,14 +94,14 @@ Display createDisplay({
     const units = ['k', 'M', 'G', 'T', 'P'];
     final unit = units[sections.length - 2];
     space = length - negative.length - mainSection.length - 1;
-    if (space >= 2) {
-      final tailShow = tailSection.substring(0, min(space - 1, tailSection.length)).replaceAll(RegExp(r'0+$'), '');
-      return '${negative}${mainSection}${tailShow == '' ? '' : '.'}${tailShow}${unit}';
-    }
     if (space >= 0) {
-      return '${negative}${mainSection}${unit}';
+      roundingRst = _rounding(mainSection, tailSection, space - 1, roundingType);
+      final main = roundingRst[0];
+      final tail = roundingRst[1].replaceAll(RegExp(r'0+$'), '');
+      return '${negative}${mainSection}${tail == '' ? '' : '.'}${tail}${unit}';
     }
   }
 
-  throw Exception('length: ${length} is too small');
+  print('number_display: length: ${length} is too small for ${value}');
+  return value.toString();
 };