+++ CHANGELOG.md
@@ -18,3 +18,21 @@
 - Optimize performance.
 - Add unit test.
 - Remove decimal trailing zeros.
+
+## 2.0.1
+
+**2019-07-31**
+
+- Enlarge SDK requirement to ">=2.1.0 <3.0.0".
+
+## 2.1.0
+
+**2019-12-05**
+
+- Add feature: roundingType. Now you can set the way to round the decimal in 'round', 'floor' or 'ceil', witch default to 'round'.
+
+- When the length is too small, number-display will return the origin value as a string, instead of throwing an error.
+
+- Change param decimal name to 'precision', and default to equal to param length, witch means no additional limit.
+
+- Change param comma name to 'separator'.